% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shay_script.R
\name{GetBiggestVenues}
\alias{GetBiggestVenues}
\title{Get Largest Venues by Country}
\usage{
GetBiggestVenues(country_name, top, api_key)
}
\arguments{
\item{country_name}{The name of the country for which to retrieve venue information.}

\item{top}{The number of top venues to retrieve.}

\item{api_key}{The API key for making the API call.}
}
\value{
A ggplot object representing the capacity of the venues.
}
\description{
This function retrieves the largest venues in a specified country using an API call.
}
\examples{
GetBiggestVenues("England", 5, "your_api_key")
}
