% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matt_script.R
\name{getFootballFixtures}
\alias{getFootballFixtures}
\title{Get Football Fixtures}
\usage{
getFootballFixtures(league, season, api_key)
}
\arguments{
\item{league}{The league ID for which to retrieve fixtures.}

\item{season}{The season for which to retrieve fixtures.}

\item{api_key}{The API key for RAPID-API access.}
}
\value{
A dataframe of football fixtures including dates, teams, and scores.
}
\description{
Retrieves football fixtures for a specified league and season.
}
\examples{
\dontrun{
getFootballFixtures(39, 2020, "your_api_key")
}
}
