% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matt_script.R
\name{getFootballStandings}
\alias{getFootballStandings}
\title{Get Football Standings}
\usage{
getFootballStandings(league, season, api_key)
}
\arguments{
\item{league}{The league ID for which to retrieve standings.}

\item{season}{The season for which to retrieve standings.}

\item{api_key}{The API key for RAPID-API access.}
}
\value{
A dataframe of league standings including team names, ranks, and points.
}
\description{
Retrieves the standings for a specified league and season.
}
\examples{
\dontrun{
getFootballStandings(39, 2020, "your_api_key")
}
}
