% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerui_script.R
\name{getLeagueInfo}
\alias{getLeagueInfo}
\title{Retrieve League Information by Name}
\usage{
getLeagueInfo(league, apikey)
}
\arguments{
\item{league}{A string specifying the name of the league to be queried.}

\item{apikey}{A string representing the API key required for authentication
with the API Sports Football API.}
}
\value{
A dataframe containing the league information if the request is
successful. Each row represents a league, and columns include details
such as league id, name, and other relevant information provided by the API.
}
\description{
This function queries the API Sports Football API to retrieve information
about a football league by its given name. It constructs a query, sends a
GET request to the API, and returns a dataframe with league details if
the request is successful. Users can use this function to obtain league ID
for parameter in other functions.
}
\examples{
getLeagueInfo("Premier League", apikey)

}
