% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerui_script.R
\name{getTeamId}
\alias{getTeamId}
\title{Retrieve Team ID by Name}
\usage{
getTeamId(team, apikey)
}
\arguments{
\item{team}{A string specifying the name of the team to be queried.}

\item{apikey}{A string representing the API key required for authentication
with the API Sports Football API.}
}
\value{
An integer representing the team ID if the request is successful.
}
\description{
This function queries the API Sports Football API to retrieve the ID
of a football team by its name. It sends a GET request with the team name
as a query parameter and returns the team ID if the request is successful.
This is a helper function for other functions taking team Id as parameter.
}
\examples{
getTeamId("Manchester United", apikey)

}
