% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matt_script.R
\name{getTeamTransfers}
\alias{getTeamTransfers}
\title{Get Team Transfers}
\usage{
getTeamTransfers(team, api_key)
}
\arguments{
\item{team}{The team ID for which to retrieve transfer information.}

\item{api_key}{The API key for RAPID-API access.}
}
\value{
A dataframe of team transfers including player names, transfer dates, and transfer details.
}
\description{
Retrieves transfer information for a specified team.
}
\examples{
\dontrun{
getTeamTransfers(33, "your_api_key")
}
}
