% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matt_script.R
\name{makeApiRequest}
\alias{makeApiRequest}
\title{Make API Request}
\usage{
makeApiRequest(url, queryString, api_key)
}
\arguments{
\item{url}{The API endpoint URL.}

\item{queryString}{A list of query parameters for the API request.}

\item{api_key}{The API key for authenticating the request to RAPID-API.}
}
\value{
A dataframe with the response data.
}
\description{
This helper function sends a GET request to a specified URL with given query parameters and headers,
then processes the response into a dataframe.
}
\examples{
\dontrun{
makeApiRequest("https://api-football-v1.p.rapidapi.com/v3/leagues", list(country="England"), "your_api_key")
}
}
