% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerui_script.R
\name{searchPlayer}
\alias{searchPlayer}
\title{Search Player Information}
\usage{
searchPlayer(name, team, apikey)
}
\arguments{
\item{name}{The name of the player to search for.}

\item{team}{The name of the team the player is associated with.}

\item{apikey}{Your API key for accessing the API-Sports API.}
}
\value{
A \code{gt} table containing detailed information about the player.
If the request fails, the function stops with an error message.
}
\description{
Searches for information about a player by name and team using the API-Sports API.
It retrieves the player's detailed information and displays it in a \code{gt} table format.
}
\examples{
searchPlayer("Neymar", "Paris Saint Germain", "<your_api_key>")
}
