% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shay_script.R
\name{GetBiggestVenues}
\alias{GetBiggestVenues}
\title{Get Largest Venues by Country}
\usage{
GetBiggestVenues(country_name, top, api_key)
}
\arguments{
\item{country_name}{The name of the country for which to retrieve venue information.}

\item{top}{The number of top venues to retrieve.}

\item{api_key}{The API key for making the API call or "mock_api_key" for using mock data.}
}
\value{
A ggplot object visualizing the capacities of the largest venues.
}
\description{
Retrieves the largest venues in a specified country. This function can
operate with actual API calls or use mock data for testing purposes when
supplied with "mock_api_key" as the API key.
}
\examples{
# Using mock API key for demonstration; replace "mock_api_key" with your actual API key.
GetBiggestVenues("England", 5, "mock_api_key")
}
