% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shay_script.R
\name{GetCoachInfo}
\alias{GetCoachInfo}
\title{Get Coach Information for a Team}
\usage{
GetCoachInfo(team_id, api_key)
}
\arguments{
\item{team_id}{The ID of the team for which to fetch coach information.}

\item{api_key}{The API key for accessing the data.}
}
\value{
A data frame containing coach information.
}
\description{
Fetches information about the coaches for a specified team. This function can
use actual API calls.
}
\examples{
GetCoachInfo("team_id_example", "api_key")
}
